Draft how rust code could look like with interthread message passing.
For any object created of the context, we use handles because having to return actual addresses for things that will be created by the domain thread (like a tweet stream), thus asynchronously, is very tricky. Better use abstract handles that we can "allocate" on the UI thread and pass to the domain thread to associate anything it needs with it.